Former-commit-id: 83399f644af46b937dd5ae7b86bf88a3ab1f4a6f

+++ ifs-data-service/src/main/java/org/innovateuk/ifs/invite/repository/CompetitionInviteRepository.java
@@ -20,6 +20,6 @@ public interface CompetitionInviteRepository extends CrudRepository<CompetitionI
 
     CompetitionInvite getByHash(String hash);
 
-    long countByCompetitionIdAndStatusIn(long competitionId, Set<InviteStatus> statuses);
+    int countByCompetitionIdAndStatusIn(long competitionId, Set<InviteStatus> statuses);
 
 }
